% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert_colors}
\alias{convert_colors}
\title{Convert colors between colorspaces}
\usage{
convert_colors(colors, from, to)
}
\arguments{
\item{colors}{A matrix of colors}

\item{from}{The colorspace of the input colors, one of "rgb", "hsl",
"din99d", "lab", "xyz"}

\item{to}{The colorspace of the output colors,
one of "rgb", "hsl", "lab", "xyz"}
}
\value{
The colors converted to the new colorspace
}
\description{
Convert colors between colorspaces
}
