% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette.R
\name{get_palette}
\alias{get_palette}
\title{Retrieve one of the built-in color palettes}
\usage{
get_palette(palette = "ColorBrewer:Accent")
}
\arguments{
\item{palette}{A character string specifying the name of the color palette to}
}
\value{
A character vector of colors in hex format.
}
\description{
This function retrieves a color palette from the built-in palettes. To
see the available palettes, use \code{\link[=list_palettes]{list_palettes()}}.
}
\examples{
get_palette("Vermeer:LittleStreet")

}
\seealso{
\code{\link[=list_palettes]{list_palettes()}}
}
