% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_palettes.R
\name{list_palettes}
\alias{list_palettes}
\title{List available color palettes}
\usage{
list_palettes()
}
\value{
A list of available color palettes. Each palette is a named list
with a character vector.
}
\description{
List available color palettes
}
\examples{
list_palettes()

}
\seealso{
\code{\link[=qualpal]{qualpal()}}
}
