% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.qualpal}
\alias{plot.qualpal}
\title{Demos and diagnostic plots of qualitative color palettes}
\usage{
\method{plot}{qualpal}(qualpal)
}
\arguments{
\item{qualpal}{A list object of class \code{"qualpal"} generated from
\code{\link{qualpal}}.}
}
\description{
Uses the colors from a \code{qualpal} object to plot \enumerate{\item a
grouped scatter plot (\code{kmeans} on \code{mtcars}), \item a heatmap of the
correlation matrix given by the the Delta-E CIEDE2000 difference computations
(darker greys mean more correlated), \item a multidimensional scaling (MDS)
map computed by \code{cmdscale} from the Delta-E distance matrix, and \item a
\code{pairs} plot of the hsl color space dimensions. }
}
\examples{
col_pal <- qualpal(n = 8)
plot(col_pal)
}

