% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predefined_colorspaces}
\alias{predefined_colorspaces}
\title{Predefined colorspaces}
\usage{
predefined_colorspaces(colorspace)
}
\arguments{
\item{colorspace}{Character vector}
}
\value{
Specification for HSL color space as a list.
}
\description{
Predefined colorspaces
}
\examples{
NULL
}
\keyword{internal}
