% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.qualpal}
\alias{print.qualpal}
\title{Print qualitative color palettes}
\usage{
\method{print}{qualpal}(x, colorspace = c("HSL", "DIN99d", "RGB"),
  digits = 2, ...)
}
\arguments{
\item{x}{A n object of class \code{"qualpal"}.}

\item{colorspace}{Colorspace to print colors in}

\item{digits}{Number of significant digits to print.
(See \link{print.default}.) Setting it to \code{NULL} uses
\code{\link{getOption}("digits")}.}

\item{\dots}{}
}
\value{
Prints the colors as a matrix in the specified color space as well
  as a distance matrix of the color differences. Invisibly returns x.
}
\description{
Print qualitative color palettes
}
\examples{
f <- qualpal(3)
print(f, colorspace = "DIN99d", digits = 3)
}
