% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopal.R, R/qualpal-deprecated.R
\name{autopal}
\alias{autopal}
\alias{qualpal-deprecated}
\title{Deprecated functions in package \pkg{qualpal}.

The functions listed below are deprecated and will be defunct in the next
major version future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.}
\usage{
autopal(
  n,
  colorspace = list(h = c(0, 360), s = c(0.2, 0.5), l = c(0.6, 0.85)),
  cvd = c("protan", "deutan", "tritan"),
  target = 20
)
}
\description{
Deprecated functions in package \pkg{qualpal}.

The functions listed below are deprecated and will be defunct in the next
major version future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.
}
\section{\code{autopal}}{

This function is deprecated and will be removed in a future release.
No direct replacement is available.
}

\keyword{internal}
