% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualpalr.R
\docType{package}
\name{qualpalr-package}
\alias{qualpalr}
\alias{qualpalr-package}
\title{qualpalr: Automatic Generation of Qualitative Color Palettes}
\description{
Automatic generation of maximally distinct qualitative color palettes, optionally tailored to color deficiency. A set of colors or a subspace of a color space is used as input and a final palette of specified size is generated by picking colors that maximize the minimum pairwise difference among the chosen colors. Adaptations to color vision deficiency, background colors, and white points are supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jolars.github.io/qualpalr/}
  \item Report bugs at \url{https://github.com/jolars/qualpalr/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johan@jolars.co} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

}
\keyword{internal}
