% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_palette.R
\name{analyze_palette}
\alias{analyze_palette}
\title{Analyze a categorical color palette}
\usage{
analyze_palette(
  palette,
  cvd = c("protan", "deutan", "tritan"),
  cvd_severity = 1,
  bg = NULL,
  metric = c("din99d", "ciede2000", "cie76")
)
}
\arguments{
\item{palette}{Either a matrix of RGB values (with values between 0 and 1),
a data frame with RGB values, or a character vector of hex colors.}

\item{cvd}{Color vision deficiency adaptation. Use \code{cvd_severity}
to set the severity of color vision deficiency to adapt to. Permissible
values are \code{"protan", "deutan",} and \code{"tritan"}.}

\item{cvd_severity}{Severity of color vision deficiency to adapt to. Can take
any value from 0, for normal vision (the default), and 1, for dichromatic
vision.}

\item{bg}{Backgrouind color to consider (but not include) when
generating the palette. This is useful to avoid colros that are too
close to the background/canvas color. If \code{NULL} (the default), the
background color is not considered at all. Any color that is convertable
via \code{\link[grDevices]{col2rgb}} is acceptable, including hex colors.}

\item{metric}{The color metric to use for the color distance
matrix.}
}
\value{
A list of lists, one for each type of color vision deficiency plus
normal vision. Each list contains \code{difference_matrix}, \code{min_distances}, and
\code{bg_min_distance}
}
\description{
Analyze a categorical color palette with respect to
the differences between the colors in the palette.
}
\examples{
pal <- qualpal(5)
analyze_palette(pal$hex, cvd = "protan", cvd_severity = 0.5)

}
\seealso{
\code{\link[=qualpal]{qualpal()}}
}
