% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.qualpal}
\alias{plot.qualpal}
\title{Multidimensional scaling map of qualitative color palette}
\usage{
\method{plot}{qualpal}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"qualpal"} generated from
\code{\link{qualpal}}.}

\item{\dots}{Arguments to pass on to \code{\link[graphics]{plot}}.}
}
\description{
Uses the colors in a \link{qualpal} object to compute and plot a
multidimensional scaling (MDS) map using \code{\link[stats]{cmdscale}} on the
Delta E DIN99d distance matrix.
}
\examples{
col_pal <- qualpal(3)
plot(col_pal)
}
\seealso{
\code{\link{qualpal}}, \code{\link{pairs.qualpal}},
  \code{\link[graphics]{plot}}
}
