% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualpal.R
\name{print.qualpal}
\alias{print.qualpal}
\title{Print qualpal palette}
\usage{
\method{print}{qualpal}(x, colorspace = c("HSL", "DIN99d", "RGB"),
  digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{"qualpal"}.}

\item{colorspace}{Color space to print colors in.}

\item{digits}{Number of significant digits for the output.
(See \link{print.default}.) Setting it to \code{NULL} uses
\code{\link{getOption}("digits")}.}

\item{\dots}{Arguments to pass to \code{\link{print.default}}.}
}
\value{
Prints the colors as a matrix in the specified color space as well
  as a distance matrix of the color differences. Invisibly returns x.
}
\description{
Print the result from a call to \code{\link{qualpal}}.
}
\examples{
f <- qualpal(3)
print(f, colorspace = "DIN99d", digits = 3)
}
