% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualpalr.R
\docType{package}
\name{qualpalr-package}
\alias{qualpalr}
\alias{qualpalr-package}
\title{qualpalr: Automatic Generation of Qualitative Color Palettes}
\description{
Automatic generation of maximally distinct qualitative color palettes, optionally tailored to color deficiency. A list of colors or a subspace of a color space is used as input and then projected to the DIN99d color space, where colors that are maximally distinct are chosen algorithmically.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jolars.github.io/qualpalr/}
  \item Report bugs at \url{https://github.com/jolars/qualpalr/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

}
\keyword{internal}
