% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualpalr.R
\docType{package}
\name{qualpalr-package}
\alias{qualpalr}
\alias{qualpalr-package}
\title{qualpalr: Automatic Generation of Qualitative Color Palettes}
\description{
Automatic generation of distinct qualitative color palettes,
    optionally adapted to color deficiency. It takes a subspace of the HSL color
    space as input and projects it to the DIN99d color space where it selects
    and return colors that are maximally distinct.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/qualpalr}
  \item Report bugs at \url{https://github.com/jolars/qualpalr/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{mail@larssonjohan.com} (0000-0002-4029-5945)

}
\keyword{internal}
